exercise 4.21

+++ part4/bloglist/controllers/blogs.js
@@ -42,6 +42,14 @@ blogsRouter.post('/', async (request, response, next) => {
 });
 
 blogsRouter.delete('/:id', async (request, response, next) => {
+  const decodedToken = jwt.verify(request.token, process.env.SECRET);
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token invalid' });
+  }
+  const blog = await Blog.findById(request.params.id);
+  if (!(blog.user.toString() === decodedToken.id)) {
+    return response.status(401).json({ error: 'token invalid' });
+  }
   try {
     await Blog.findByIdAndDelete(request.params.id);
     response.status(204).end();
